rag_api/pyproject.toml
[project]
name = "rag_api"
version = "0.1.0"
description = "RAG API service for document querying"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "openai>=1.12.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.1",
    "httpx==0.27.2",
    "langchain>=0.1.9",
    "langchain-openai>=0.0.5"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"] 